% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flights-example.R
\docType{data}
\name{flights_example}
\alias{flights_example}
\title{Subset of nycflights13 dataset}
\format{An object of class \code{data.frame} with 118 rows and 4 columns.}
\source{
\url{https://github.com/hadley/nycflights13}
}
\usage{
flights_example
}
\description{
A subset of the nycflights13 dataset by Hadley Wickham.
}
\keyword{datasets}
