% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-domino.R
\name{geom_domino}
\alias{geom_domino}
\title{Angled line segments between x-axis values}
\usage{
geom_domino(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
\code{geom_domino} draws a sloping line between points (x, y) and
(xend, y), where y is an arbitrary height above x and below xend.  If x
equals xend then the line is vertical.
}
\details{
This is useful for visualising the difference in time between two events
related to one object, e.g. scheduled and actual arrival.
}
\section{Aesthetics}{


{\code{geom_segment} understands the following aesthetics (required aesthetics are in bold):

\itemize{
  \item \code{\strong{x}}
  \item \code{\strong{y}}
  \item \code{\strong{xend}}
  \item \code{\strong{yend}}
  \item \code{alpha}
  \item \code{colour}
  \item \code{group}
  \item \code{linetype}
  \item \code{size}
}
}

}

\examples{
# Plot a subset of the nycflights13 data
library(ggplot2)
ggplot(ggdomino::flights_example,
         aes(sched_dep_time, carrier, xend = dep_time, colour = late)) +
  geom_domino() +
  scale_colour_manual("Late", values = c("black", "red")) +
  xlab("") +
  ylab("Carrier") +
  ggtitle("Departure times from New York to Boston", sub = "Scheduled vs actual") +
  theme_minimal() +
  theme(panel.grid = element_blank())
}
\seealso{
\code{\link{geom_segment}}, from which \code{geom_domino} inherits.
}
