% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timetable.R
\docType{data}
\name{timetable}
\alias{timetable}
\title{Part of the Metropolitan Line timetable}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 236 rows and 5 columns.}
\source{
\url{https://api.tfl.gov.uk/}
}
\usage{
timetable
}
\description{
A subset of the timetable of the London Underground Metropolitan Line,
travelling from Harrow-on-the-Hill to Moor Park.  Works best when the 'hms'
package is loaded.
}
\keyword{datasets}
